Supporting new debugger

+++ vnext/ReactCommon/TEMP_UntilReactCommonUpdate/jsinspector-modern/NetworkIOAgent.h
@@ -34,8 +34,8 @@ struct LoadNetworkResourceRequest {
 
 struct ReadStreamParams {
   StreamID handle;
-  std::optional<unsigned long> size;
-  std::optional<unsigned long> offset;
+  std::optional<size_t> size; // [Windows #13587]
+  std::optional<size_t> offset; // [Windows #13587]
 };
 
 struct NetworkResource {
@@ -236,7 +236,7 @@ class NetworkIOAgent {
    * Stream IDs are strings of an incrementing integer, unique within each
    * NewtworkIOAgent instance. This stores the next one to use.
    */
-  unsigned long nextStreamId_{0};
+  size_t nextStreamId_{0}; // [Windows #13587]
 
   /**
    * Begin loading an HTTP resource, delegating platform-specific
